+++ src/app/components/reports/reportsController.js
@@ -1,8 +1,8 @@
 'use strict';
 
 angular.module('liveopsConfigPanel')
-  .controller('ReportsController', ['$scope', '$sce', '$http', 'Session', 'Report', '$state', 'BIRST_URL', '$timeout', '$translate',
-    function ($scope, $sce, $http, Session, Report, $state, BIRST_URL, $timeout, $translate) {
+  .controller('ReportsController', ['$scope', '$sce', '$http', 'Session', 'Report', '$state',  'BIRST_URL', '$timeout', '$translate', 'appFlags',
+    function ($scope, $sce, $http, Session, Report, $state, BIRST_URL, $timeout, $translate, appFlags) {
       $scope.birst = {};
       $scope.dashboardReady = false;
       $scope.birst.message = $translate.instant('reports.default');
@@ -15,9 +15,16 @@ angular.module('liveopsConfigPanel')
         $scope.dashboardReady = true;
         $scope.$apply();
 
+        // feature flag to temporarily suppress click handler that
+        // open the app dock
+        if (!appFlags.APPDOCK) {
+          return;
+        } 
+
         document.domain = 'cxengagelabs.net';
         var js = "function interceptClickEvent(e){e.preventDefault();var hrefPath;var target=e.target||e.srcElement;if(target.tagName==='A'){hrefPath=target.getAttribute('href');if(hrefPath.startsWith('https://parent.cxengagelabs.net/')){parent.addApp({type:'recording',id:hrefPath})}}e.preventDefault()} $(document).on('click',interceptClickEvent);";
-          frames[0].window.eval(js);
+
+        frames[0].window.eval(js);
 
       });