+++ src/app/shared/directives/appDock/appDockDirective.js
@@ -1,8 +1,8 @@
 'use strict';
 
 angular.module('liveopsConfigPanel')
-  .directive('appDock', ['$window', '$rootScope', '$timeout',
-    function ($window, $rootScope, $timeout) {
+  .directive('appDock', ['$window', '$rootScope', '$timeout', 'appFlags',
+    function ($window, $rootScope, $timeout, appFlags) {
       return {
         restrict: 'E',
         templateUrl: 'app/shared/directives/appDock/appDock.html',
@@ -11,6 +11,11 @@ angular.module('liveopsConfigPanel')
           toolbarStatus: '='
         },
         link: function (scope, attrs, elem, controller) {
+          // feature flag to temporarily suppress the appdock
+          if (!appFlags.APPDOCK) {
+            return;
+          }
+
           scope.apps = [];
           scope.showAppDock = false;
           scope.showApps = false;